backend/requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
rapidfuzz==3.9.6
pdfplumber==0.11.4
fpdf==1.7.2
dataframe-image==0.2.4
html2image==2.0.4
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
openpyxl==3.1.5
